Volatile memory > Quadrant threading
Threw out the idea of quadrant threading for volatile images, which sped
up background frame drawing by 3x.
Background frame is still a BufferedImage, but each frame a volatile
image is created and is directly drawn to, (the individual
polygon-drawing is proportional to particle quantity and size) then that
v-Image is drawn onto the b-image which is later grabbed by the panel
and drawn to screen.